+++ atlasdb-client/src/main/java/com/palantir/atlasdb/AtlasDbMetricNames.java
@@ -51,6 +51,7 @@ public final class AtlasDbMetricNames {
     public static final String PREFIX_CONSERVATIVE = "conservative";
     public static final String PREFIX_THOROUGH = "thorough";
     public static final String ENQUEUED_WRITES = "enqueuedWrites";
+    public static final String ENTRIES_READ = "entriesRead";
     public static final String TOMBSTONES_PUT = "tombstonesPut";
     public static final String ABORTED_WRITES_DELETED = "abortedWritesDeleted";
     public static final String SWEEP_TS = "sweepTimestamp";