+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/sweep/queue/ShardAndStrategy.java
@@ -32,6 +32,10 @@ public abstract class ShardAndStrategy {
                 + "THOROUGH, but it is %s instead.", strategy());
     }
 
+    public String toText() {
+        return "shard " + shard() + " and strategy " + strategy();
+    }
+
     public boolean isConservative() {
         return strategy() == TableMetadataPersistence.SweepStrategy.CONSERVATIVE;
     }